spawnpoint stay on track
the spwanpoints moved together with the player left and right. The
reason for this behavior was that the Spawner-object was connected to
the player and the spawnpoints under the Spawner-folder. Looks like if a
GameObject is under a folder the objects in this folder have an
hardcoded position to the mother-object